requirements.txt
aiogram==2.25.1
beautifulsoup4==4.12.2
deep_translator==1.10.1
googletrans==3.0.0
Pillow==9.5.0
pytesseract==0.3.10
Requests==2.30.0
